move files to reflect class hierarchy
