+++ spec/spec_helper.rb
@@ -4,5 +4,8 @@ require_relative "../lib/lumberjack_json_device"
 
 RSpec.configure do |config|
   config.warnings = true
+  config.disable_monkey_patching!
+  config.default_formatter = "doc" if config.files_to_run.one?
   config.order = :random
+  Kernel.srand config.seed
 end
